+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+- Add TypeScript support for custom world in hooks, fixes [#1304](https://github.com/badeball/cypress-cucumber-preprocessor/issues/1304).
+
 ## v22.1.0
 
 - Gracefully handle sourcemaps when running component tests with webpack and chunks enabled (default behavior), fixes [#1296](https://github.com/badeball/cypress-cucumber-preprocessor/issues/1296).